VR: Optimize Header and update CLI interface (breaking)
This is a breaking change for client and cluster init that:

* adds `Header.head` as a u128,
* adds `Header.version` as a u8,
* resizes `Header.cluster` from u128 to u32,
* accepts `Header.cluster` now being zero (no longer a UUID),
* resizes `Header.view` from u64 to u32,
* resizes `Header.replica` from u16 to u8, and
* adds a client table fix.

The motivation behind this is to:

* add space in the header to support end-to-end integrity
  checking in the client to match replies back to requests,
  and prevent future issues with protocol versioning, and

* make the cluster CLI more memorable and easier to use:

  --cluster=0 --replica=0 --addresses=3001,3002,3003

+++ src/test_message_bus.zig
@@ -49,11 +49,11 @@ pub const MessageBus = struct {
     }
 
     /// Returns true if the target replica is connected and has space in its send queue.
-    pub fn can_send_to_replica(self: *MessageBus, replica: u16) bool {
+    pub fn can_send_to_replica(self: *MessageBus, replica: u8) bool {
         return !self.send_queues[replica].full();
     }
 
-    pub fn send_header_to_replica(self: *MessageBus, replica: u16, header: Header) void {
+    pub fn send_header_to_replica(self: *MessageBus, replica: u8, header: Header) void {
         assert(header.size == @sizeOf(Header));
 
         if (!self.can_send_to_replica(replica)) {
@@ -77,7 +77,7 @@ pub const MessageBus = struct {
         self.send_message_to_replica(replica, message);
     }
 
-    pub fn send_message_to_replica(self: *MessageBus, replica: u16, message: *Message) void {
+    pub fn send_message_to_replica(self: *MessageBus, replica: u8, message: *Message) void {
         self.send_queues[replica].push(message.ref()) catch |err| switch (err) {
             error.NoSpaceLeft => {
                 self.unref(message);
