+++ src/vr/journal.zig
@@ -18,12 +18,12 @@ pub const Journal = struct {
     pub const Read = struct {
         self: *Journal,
         completion: Storage.Read,
-        callback: fn (self: *Replica, prepare: ?*Message, destination_replica: ?u16) void,
+        callback: fn (self: *Replica, prepare: ?*Message, destination_replica: ?u8) void,
 
         message: *Message,
         op: u64,
         checksum: u128,
-        destination_replica: ?u16,
+        destination_replica: ?u8,
     };
 
     pub const Write = struct {
@@ -84,7 +84,7 @@ pub const Journal = struct {
 
     allocator: *Allocator,
     storage: *Storage,
-    replica: u16,
+    replica: u8,
     size: u64,
     size_headers: u64,
     size_circular_buffer: u64,
@@ -120,7 +120,7 @@ pub const Journal = struct {
     pub fn init(
         allocator: *Allocator,
         storage: *Storage,
-        replica: u16,
+        replica: u8,
         size: u64,
         headers_count: u32,
         init_prepare: *Header,
@@ -431,10 +431,10 @@ pub const Journal = struct {
 
     pub fn read_prepare(
         self: *Journal,
-        callback: fn (replica: *Replica, prepare: ?*Message, destination_replica: ?u16) void,
+        callback: fn (replica: *Replica, prepare: ?*Message, destination_replica: ?u8) void,
         op: u64,
         checksum: u128,
-        destination_replica: ?u16,
+        destination_replica: ?u8,
     ) void {
         const replica = @fieldParentPtr(Replica, "journal", self);
         if (op > replica.op) {
